Refactor, fix chunking, and fix VcfReader
