+++ inc/fetch.hpp
@@ -1,0 +1,53 @@
+#pragma once
+
+#include "read_cluster_optimizer.hpp"
+#include "TransitiveMap.hpp"
+#include "IntervalGraph.hpp"
+#include "Authenticator.hpp"
+#include "VcfReader.hpp"
+#include "windows.hpp"
+#include "Region.hpp"
+#include "Timer.hpp"
+#include "CLI11.hpp"
+#include "misc.hpp"
+#include "bed.hpp"
+#include "bam.hpp"
+
+#include "bindings/cpp/WFAligner.hpp"
+using namespace wfa;
+
+#include <unordered_map>
+#include <exception>
+#include <stdexcept>
+#include <iostream>
+#include <fstream>
+#include <thread>
+#include <memory>
+#include <limits>
+
+using std::numeric_limits;
+using std::unordered_map;
+using std::runtime_error;
+using std::exception;
+using std::ifstream;
+using std::ofstream;
+using std::thread;
+using std::atomic;
+using std::mutex;
+using std::cerr;
+using std::max;
+using std::min;
+using std::cref;
+using std::ref;
+
+namespace sv_merge{
+
+void fetch_reads(
+        Timer& t,
+        vector<Region>& regions,
+        path bam_csv,
+        int64_t n_threads,
+        unordered_map<Region,TransMap>& region_transmaps
+);
+
+}
