+++ src/windows.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 #include "windows.hpp"
 #include "VcfReader.hpp"
 #include "bed.hpp"
@@ -9,9 +7,9 @@ namespace sv_merge{
 
 void construct_windows_from_vcf_and_bed(path tandem_bed, path vcf, int32_t flank_length, int32_t interval_max_length, vector<Region>& regions){
     VcfReader vcf_reader(vcf);
+    vcf_reader.min_qual = numeric_limits<float>::min();
     vcf_reader.min_sv_length = 0;
     vcf_reader.progress_n_lines = 100'000;
-    vcf_reader.min_sv_length = 20;
 
     unordered_set<uint32_t> sample_ids;
     unordered_set<string> sample_names;
@@ -38,7 +36,7 @@ void construct_windows_from_vcf_and_bed(path tandem_bed, path vcf, int32_t flank
             return;
         }
 
-        r.get_reference_coordinates(true, coord);
+        r.get_reference_coordinates(false, coord);
 
         contig_intervals[r.chrom].emplace_back(coord, false);
     });
@@ -49,22 +47,26 @@ void construct_windows_from_vcf_and_bed(path tandem_bed, path vcf, int32_t flank
     for (auto& [contig, intervals]: contig_intervals){
         cerr << "\tStarting: " << contig << '\n';
 
+        if (contig != "chr20"){
+            continue;
+        }
+
         vector<interval_t> components;
 
-        get_overlapping_components(intervals, components);
+        get_overlapping_components(flank_length, intervals, components);
 
-        g.for_each_connected_component_interval([&](interval_t& interval, unordered_set<string>& values){
-            if (interval.second - interval.first > interval_max_length){
-                return;
+        for (const auto& c: components){
+            if (c.second - c.first > interval_max_length){
+                continue;
             }
 
-            regions.emplace_back(contig, interval.first, interval.second - flank_length);
-        });
+            regions.emplace_back(contig, c.first, c.second);
+        }
     }
 }
 
 
-void get_overlapping_components(vector <pair <interval_t, bool> >& labeled_intervals, vector <interval_t>& result, int32_t min_gap_length){
+void get_overlapping_components(int32_t min_gap_length, vector <pair <interval_t, bool> >& labeled_intervals, vector <interval_t>& result){
     // How to sort labeled intervals
     auto left_comparator = [](const pair <interval_t, bool>& a, const pair <interval_t, bool>& b){
         return a.first.first < b.first.first;
@@ -75,7 +77,7 @@ void get_overlapping_components(vector <pair <interval_t, bool> >& labeled_inter
     result.clear();
     result.emplace_back(labeled_intervals.front().first);
 
-    bool has_non_tandem_interval;
+    bool has_non_tandem_interval = false;
 
     for (auto& [interval,is_tandem] : labeled_intervals){
         if (interval.first > interval.second){
@@ -93,12 +95,9 @@ void get_overlapping_components(vector <pair <interval_t, bool> >& labeled_inter
             else{
                 // Otherwise, start a new interval
                 result.emplace_back(interval);
-
-                // Trim the buffer off the completed component
-                result.back().second -= min_gap_length;
             }
 
-            has_non_tandem_interval = false;
+            has_non_tandem_interval = not is_tandem;
         }
 
         // Maintain the maximum bound of this component
@@ -114,7 +113,13 @@ void get_overlapping_components(vector <pair <interval_t, bool> >& labeled_inter
         // Return the stored interval to original state
         interval.second -= min_gap_length;
     }
+
+    for (auto& item: result){
+        item.second -= min_gap_length;
+    }
+
 }
+
 
 
 }
